refactor(pdg): improve interface of var_context
`build_expression` previously returned a `TemplateRecord` because
it's called internally in `VarContext` and a TR is needed there.
However, these TRs are never necessary outside of the class, where
they're only used to access the contained data node. Therefore, we'll
have `build_expression` return the data node directly, and have a
private method `_build_expression` to return the TR.