refactor(pdg): improve handling of literal/broken exprs
Previously we had to use a hack when template expressions were in
fact literals or malformed, because the expected return type was a
template record, which requires an expression node that is unused
for literals. Now that the API was split, and since `VarContext`
itself never provides broken expressions, we can move this check
to the external API, which is cleaner.