refactor(pdg): switch to dataclasses in var records
We were already using named tuples for 2 of the 3, but not for the
last one because named tuples don't mix well with inheritance.
Dataclasses do allow inheritance, and generate equality and hash
methods for us. This will eventually allow us to store the records
in the var context dicts and in other records.