feat(pdg): support for composite data
Two aspects are left unsupported for now:
    * Computed dictionary keys (expressions in attribute names)
    * Composite data in conditions

The former could be fixed by adding a special-purpose node to
aggregate the key and value data nodes, or switching to a special
graph that allows us to put edges ending in edges to denote the
usage of the value in the keyword edge.
The latter may be addressed by more refactoring, but likely doesn't
happen often.

Existing limitation: Re-using a composite value of which at least one
constituent has an impure/changed expression, while another constituent
is a literal, will cause the literal to be duplicated even though it's
the same in both usages, and occurs in the same location. This isn't
necessarily wrong, but sub-optimal. A potential fix could look into
implementing a store mapping from literal value and source location to
the literal node, so that we'll always reuse the same node for the
same literals.