feat(pdg): explicitly store control dependences
Previously there was an indirection in control dependences, i.e.,
the Conditional and Loop nodes. Now, those nodes no longer exist,
and the data governing the condition or loop is instead linked directly
to the iterated/conditionally executed tasks.

This both resembles a real PDG more closely, and improves the mineability
of the graphs.