feat: migrate to Kuzu as graph database
Requiring a Redis instance to be running in the background is not
very user-friendly, and RedisGraph is deprecated and has reached
end-of-life. Therefore, we're switching to Kuzu, an embedded, in-memory
graph database with Cypher support.

To support this, the graph queries were rewritten to work around some of
the incompatibilities in Cypher support between Kuzu and RedisGraph. We
can now make use of regular expressions again, leading to more reliable
and comprehensible pattern matching instead of string contains tests.
Moreover, subqueries are supported, simplifying some of the smell checks.

Initial experimentation suggests that Kuzu appears to be quite a bit slower
than RedisGraph though, but the improved practicality outweighs this performance
hit.